Introduce the abstract `MessageAction` base for Java
